so much refactoring and optimizations

+++ src/core/forked.ts
@@ -1,11 +1,18 @@
 import { ChildProcess, spawn } from 'child_process';
 import { PidmanSysUtils, PidmanProcessUtils } from '../utils';
 import { ProcessOptions } from './process';
+import { get } from 'lodash';
 
 let forkedProcess: ForkedProcess;
 
+const catchError = function (error): void {
+  console.error(error);
+  process.exit();
+};
+
 export enum ForkedMessageType {
   started,
+  closed,
   options,
   kill,
   killed,
@@ -35,19 +42,39 @@ class ForkedProcess {
       cwd: this.options.path,
       env: this.options.envVars || {},
       gid: PidmanSysUtils.getGid(this.options.group || ''),
-      shell: this.options.shell || false,
-      // detached: true,
+      shell: false,
+      detached: true,
       stdio: 'inherit',
       windowsHide: true
     });
 
-    // this.#child.unref();
+    this.#child?.on('close', (code, signal) => {
+      if (process.send) {
+        process.send(
+          new ForkedMessage(ForkedMessageType.closed, {
+            exitCode: code, exitSignal: signal
+          })
+        )
+      }
+    });
+    this.#child?.on('error', catchError);
+    this.#child.stderr?.on('data', catchError);
+
+    this.#child.unref();
 
     return this.#child.pid;
   }
 
   getPid(): number {
     return this.#child!.pid;
+  }
+
+  getChild(): ChildProcess {
+    return this.#child!;
+  }
+
+  getOptions(): ProcessOptions {
+    return this.options;
   }
 }
 
@@ -70,17 +97,23 @@ process.on('message', async (msg: ForkedMessage) => {
     // kill request
     if (msg.type === ForkedMessageType.kill) {
       try {
-        const childrens = await PidmanProcessUtils.getPidChildrens(
-          forkedProcess.getPid()
-        );
-        const killed = await PidmanProcessUtils.killTree(
-          forkedProcess.getPid()
-        );
+        let killed;
+        const pid = forkedProcess.getPid();
+        const exitCode = get(forkedProcess.getChild(), 'exitCode');
+
+        if (pid && exitCode === null) {
+          killed = await PidmanProcessUtils.killTree(
+            pid, forkedProcess.getOptions().killSignal
+          );
+        } else {
+          killed = false;
+        }
 
         process.send(
-          new ForkedMessage(ForkedMessageType.killed, childrens)
+          new ForkedMessage(ForkedMessageType.killed, killed)
         );
       } catch (err) {
+        // console.error(err);
         process.send(
           new ForkedMessage(ForkedMessageType.fail, err)
         );
